bin/getchangelog.py: Include short package description in Google+ output
It has been requested that the post on Google+ does not only
tell the name, version and changelog of the application but
also (at least) a short description of what the application
is actually.
So the getchangelog.py gets an optional -d/--description parameter
which tries to read a control file in the same directory as the
changelog file. It searches for a package block with the name
of the source package and extracts the description out of it.