formatChangelog.py: Fix endless loop with too long line
In case a line has a very long text with no spaces in it (e.g. an URL)
it tries forever to find a space where it can break the line.
So also just print the warning when the rest of the line is longer than
allowed and does not contain a space.