build/process_pre_build.py: Add not installed files in the mail
There is the dh_install --fail-missing flag to prevent files from
not being installed to anywhere. Unfortunately when
there are both all and any packages then the all packages
do not get installed on amd64 which then makes the build fail.
So just use --list-missing flag and list the files in the mail.